+++ src/main/java/ru/yandex/practicum/filmorate/controllers/MpaController.java
@@ -23,7 +23,7 @@ public class MpaController {
     }
 
     @GetMapping("/{id}")
-    public Mpa getById(@PathVariable int id) {
+    public Mpa getById(@PathVariable Long id) {
         return mpaService.getById(id);
     }
 }
