+++ src/main/java/ru/yandex/practicum/filmorate/storage/mapper/GenreMapper.java
@@ -12,7 +12,7 @@ public class GenreMapper implements RowMapper<Genre> {
     @Override
     public Genre mapRow(ResultSet rs, int rowNum) throws SQLException {
         return Genre.builder()
-                .id(rs.getInt("genre_id"))
+                .id(rs.getLong("genre_id"))
                 .name(rs.getString("name"))
                 .build();
     }
