+++ src/main/java/ru/yandex/practicum/filmorate/storage/mpa/MpaStorage.java
@@ -8,5 +8,5 @@ public interface MpaStorage {
 
     Collection<Mpa> findAll();
 
-    Mpa getById(int id);
+    Mpa getById(Long id);
 }
