+++ src/main/java/ru/yandex/practicum/filmorate/storage/reviewsLikes/ReviewLikeStorage.java
@@ -1,11 +1,11 @@
 package ru.yandex.practicum.filmorate.storage.reviewsLikes;
 
 public interface ReviewLikeStorage {
-    void addLike(int reviewId, int userId);
+    void addLike(Long reviewId, Long userId);
 
-    void addDislike(int reviewId, int userId);
+    void addDislike(Long reviewId, Long userId);
 
-    boolean deleteLike(int reviewId, int userId);
+    boolean deleteLike(Long reviewId, Long userId);
 
-    boolean deleteDislike(int reviewId, int userId);
+    boolean deleteDislike(Long reviewId, Long userId);
 }
